test(nocodb): unit-test CSV charset auto-fallback; extract encoder to helper
